fix: map results relative to rso
Search results are now mapped relative to rso (top down approach),
rather than relative to the "a > h3" selection (bottom up approach).

This is based on the conjecture that the
hierarchy between rso and the desired class remains consistent, while
the hierachy between the desired class and the "a > h3" selection is not
always consistent.